pyproject.toml
[project]
name = "music-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "chromadb>=1.3.5",
    "cyclopts>=4.2.5",
    "fastapi>=0.121.3",
    "laion-clap>=1.1.7",
    "pydantic-settings>=2.12.0",
    "scikit-learn>=1.5.2",
    "structlog>=25.5.0",
    "torchvision>=0.24.1",
    "uvicorn>=0.38.0",
]

[tool.pyright]
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

